+++ CHANGES.md
@@ -20,6 +20,7 @@ Changes in 3.3
 - CSS Rename .list-header to .list-nav & add .list-no-nav class in stylesheet.css
 - Add Grunt for automatic CSS & JS files concat & minify in package.json, Gruntfile.js, assets/js/ & assets/themes/WPadmin/
 - Relevance score inside bar (transparent) so value can be exported in stylesheet.css, colors.css & ListOutput.fnc.php
+- Check if Request exists before inserting in Requests.php
 
 Changes in 3.2
 --------------